requirements.txt
Django==5.1.7
requests==2.32.3
python-dotenv==1.0.1
django-cors-headers==4.7.0
openai==1.68.2
django-filter==25.1
pillow==11.1.0
beautifulsoup4==4.13.3
scikit-learn==1.6.1
markdown==3.7
flake8==7.1.2
unidecode==1.3.8
qrcode==8.0
python-pptx==1.0.2